SMARTSTOCKAI-AI/requirements.txt
ultralytics
opencv-python
pillow
pandas
